Changes to support m2022a Vivado 2023.1
What was needed was were determined by compiling the toolflow for all of the
RFSoC tutorials. Once those were completed the same changes that were required
were extend to other blocks that are not directly used in those tutorials but
follow similar patterns that are expected to also need to be changed.

Primary changes were:
  - when searching for mask type on blocks "Xilinx" has been removed from the
    name. E.g., "Xilinx Gateway In" is just "Gateway In".
  - When setting "explicit_period" for counters needed to set the value to
    "Explicit" or "Inferred from inputs" rather than "on"/"off". It is not clear
    why this was not needed for all instances of configuring a mask
    "explicit_period".
  - Setting `arith_type` to `Boolean` was not sufficient that the
    `gui_display_data` value also needed to be selected.

There seems to be a few quirks where setting these mask parameters (such as
`gui_display_data` for constants) does not have the same behavior as when
working with the mask in the GUI on the canvas. For example, setting
`gui_display_data` with a call to `set_param` will update and apply the mask
value but the callback to redraw or update the mask gui is not done and so
radio/fields are still active when changing the values programmatically